Rename the directory.
