+++ torchbenchmark/models/opacus_cifar10_dpsgd/requirements.txt
@@ -1,0 +1,4 @@
+torch
+torchvision
+opacus
+dm-haiku
